ajustando melhor responsabilidades de cada componente. Funçoes de cadastro no form de cadastro e de ediçao na tabela de funcionarios
